requirements.txt
astroid==2.1.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.2.5
django-bootstrap4==0.0.8
django-extensions==2.1.5
djangorestframework==3.10.2
entrypoints==0.3
Flask==1.0.2
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.17.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.7
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.4.2
python-dateutil==2.7.5
pytz==2019.3
pywinpty==0.5.5
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
scikit-learn==0.21.3
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.2.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-unicode-console==0.5
wrapt==1.11.1
